requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
requests==2.31.0
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
boto3==1.28.0
